requirements.txt
python-telegram-bot==21.11.1
openai
python-dotenv
psycopg[binary]
psycopg-pool
pgvector
beautifulsoup4
requests
